requirements.txt
streamlit
requests
bs4
re
google-generativeai